requirements.txt
minimalmodbus~=1.0.2
PyYAML~=5.3.1
gpiozero~=1.5.1
Flask~=1.1.2